relocate EntityAspectSpecs and type its API contract
